feat: moved metadata to headers, cleanups

+++ Robustor/Outbox/OutboxProcessor.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace Robustor;
+namespace Robustor.Outbox;
 
 public class OutboxBackgroundService(IServiceScopeFactory serviceScopeFactory) 
     : BackgroundService
@@ -11,16 +11,20 @@ public class OutboxBackgroundService(IServiceScopeFactory serviceScopeFactory)
         while (!stoppingToken.IsCancellationRequested)
         {
             await using var scope = serviceScopeFactory.CreateAsyncScope();
-            
             var repo = scope.ServiceProvider.GetRequiredService<IOutboxRepository>();
-            var producer = scope.ServiceProvider.GetRequiredService<IMessageProducer>();
-            
             var messages = (await repo.Get()).ToList();
+            
+            if (messages.Count == 0)
+                continue;
+
+            var producer = scope.ServiceProvider.GetRequiredService<IInternalMessageProducer>();
+
             foreach (var message in messages)
-                await producer.Produce(message.Topic, message.Message);
-            
+            {
+                await producer.Produce(message.Topic, Guid.NewGuid(), message.Message, stoppingToken);
+            }
+
             await repo.Delete(messages.Select(x => x.Id));
-            
             await Task.Delay(TimeSpan.FromSeconds(1), stoppingToken);
         }
     }
